index and webserver function separated from main.
